backend/requirements.txt
# FastAPI Backend Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Environment
python-dotenv>=1.0.0

# NLP Dependencies (use latest compatible versions)
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99

# API Integrations
requests>=2.31.0
httpx>=0.25.0

# Utilities
python-dateutil>=2.8.0
